+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.8.3
 
+* New table_column_options `cell_style` and `header_style`
+  Example:
+  ```
+  column :name, cell_style: {:'background-color' => 'red'}, header_style: {:'font-weight' => 'bold'}
+  ```
+
 * Added `classes` attribute to `table_column_options`.
   Example:
   ```